requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
future==0.18.2
idna==2.10
lxml==4.5.2
numpy==1.19.1
pandas==1.1.0
pygame==1.9.6
python-dateutil==2.8.1
python-whois==0.7.3
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
wincertstore==0.2
